+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+UNRELEASED
+----------
+- Add built css to npm build
+
 1.0.2
 -----
 - Add timeout for async dependencies in PayPal